Replace RedisMediaCache with async implementation supporting List<Media>
