+++ src/main/kotlin/com/download/downloaderbot/core/cache/MediaCache.kt
@@ -4,7 +4,7 @@ import com.download.downloaderbot.core.domain.Media
 import java.time.Duration
 
 interface MediaCache {
-    suspend fun getBySourceUrl(sourceUrl: String): Media?
-    suspend fun put(media: Media, ttl: Duration = Duration.ofDays(7)) // TODO: make ttl configurable
+    suspend fun get(sourceUrl: String): List<Media>?
+    suspend fun put(media: List<Media>, ttl: Duration = Duration.ofDays(7)) // TODO: make ttl configurable
     suspend fun evict(sourceUrl: String)
 }
